(fix) organize main components into relevant directories
