+++ src/components/stepThree/index.js
@@ -1,16 +1,17 @@
 import React from 'react'
-import '../assets/stylesheets/application.css';
+import '../../assets/stylesheets/application.css';
 import { Link } from 'react-router-dom'
-import { getWeb3, checkWeb3, setExistingContractParams } from '../utils/blockchainHelpers'
-import { stepTwo } from './stepTwo'
-import { getOldState, defaultCompanyStartDate, defaultCompanyEndDate, stepsAreValid, allFieldsAreValid } from '../utils/utils'
-import { StepNavigation } from './Common/StepNavigation'
-import { InputField } from './Common/InputField'
-import { InputFieldExt } from './Common/InputFieldExt'
-import { RadioInputField } from './Common/RadioInputField'
-import { CrowdsaleBlock } from './Common/CrowdsaleBlock'
-import { WhitelistInputBlock } from './Common/WhitelistInputBlock'
-import { NAVIGATION_STEPS, defaultState, VALIDATION_MESSAGES, VALIDATION_TYPES, TEXT_FIELDS, intitialStepThreeValidations } from '../utils/constants'
+import { getWeb3, checkWeb3, setExistingContractParams } from '../../utils/blockchainHelpers'
+import { stepTwo } from '../stepTwo'
+import { defaultCompanyStartDate } from './utils'
+import { getOldState, stepsAreValid, allFieldsAreValid, defaultCompanyEndDate } from '../../utils/utils'
+import { StepNavigation } from '../Common/StepNavigation'
+import { InputField } from '../Common/InputField'
+import { InputFieldExt } from '../Common/InputFieldExt'
+import { RadioInputField } from '../Common/RadioInputField'
+import { CrowdsaleBlock } from '../Common/CrowdsaleBlock'
+import { WhitelistInputBlock } from '../Common/WhitelistInputBlock'
+import { NAVIGATION_STEPS, defaultState, VALIDATION_MESSAGES, VALIDATION_TYPES, TEXT_FIELDS, intitialStepThreeValidations } from '../../utils/constants'
 const { CROWDSALE_SETUP } = NAVIGATION_STEPS
 const { EMPTY, VALID, INVALID } = VALIDATION_TYPES
 const { START_TIME, END_TIME, MINCAP, RATE, SUPPLY, WALLET_ADDRESS, CROWDSALE_SETUP_NAME, ALLOWMODIFYING, DISABLEWHITELISTING } = TEXT_FIELDS
